+++ test/components/Select.js
@@ -171,7 +171,7 @@ tape('Select', function (tape) {
     tape.strictEqual(
       new Select({
         type: Country,
-        options: {error: 'myerror'},
+        options: {error: 'myerror', hasError: true},
         ctx: ctx
       }).getLocals().error,
       'myerror',
@@ -183,7 +183,7 @@ tape('Select', function (tape) {
         options: {
             error: function (value) {
               return 'error: ' + value;
-            }
+            }, hasError: true
         },
         ctx: ctx,
         value: 'a'