+++ test/components/Textbox.js
@@ -408,7 +408,7 @@ tape('Textbox', function (tape) {
     tape.strictEqual(
       new Textbox({
         type: t.Str,
-        options: {error: 'myerror'},
+        options: {error: 'myerror', hasError: true},
         ctx: ctx
       }).getLocals().error,
       'myerror',
@@ -420,7 +420,7 @@ tape('Textbox', function (tape) {
         options: {
           error: function (value) {
             return 'error: ' + value;
-          }
+          }, hasError: true
         },
         ctx: ctx,
         value: 'a'